Bugfix: fixing role "RspecSkeleton" does not exist error.
If you try to create a dabase with

   rake db:create

You may see this error

   Couldn't create database for {"adapter"=>"postgresql", "encoding"=>"unicode", "database"=>"RailsRspecSkeleton_development", "pool"=>5, "username"=>"RailsRspecSkeleton", "password"=>nil}
   FATAL:  role "RailsRspecSkeleton" does not exist

I don't know what causes this error. I do know how to fix it.

You go into the database.yml file adn delete the username.

What I think is going on:

1. The Postgress database server can be used by multiple users who create multiple databases.
2. My postgres database has one root user: me.  I don't know how to create another user
3. When you try to create a database with this app, it looks for a username called RailsRspecSkeleton
4. You don't have a username called RailsRspecSkeleton, because you don't know how to create a user
5. Solution: delete the username.  This means that you, personally, will own all of the databases on your machine.

This is probably a bad thing and you should learn how to create a separate postgres user for each app.  In the meantime,
you now personally are the creator of this database.